Document changes and use SlashCommandContext builders.
