+++ src/main/kotlin/tv/blademaker/slash/context/ContextCreator.kt
@@ -1,7 +1,6 @@
 package tv.blademaker.slash.context
 
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent
-import tv.blademaker.slash.context.impl.GuildSlashCommandContext
 
 interface ContextCreator {
