+++ src/main/kotlin/tv/blademaker/slash/context/impl/GuildSlashCommandContextImpl.kt
@@ -1,0 +1,11 @@
+package tv.blademaker.slash.context.impl
+
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent
+import tv.blademaker.slash.context.GuildSlashCommandContext
+import java.util.concurrent.atomic.AtomicReference
+
+class GuildSlashCommandContextImpl(
+    override val event: SlashCommandInteractionEvent
+) : GuildSlashCommandContext {
+    override var extra: AtomicReference<Any?> = AtomicReference()
+}
